Juntando propriedades do HealthStoreService
Agora deve estar funcionando o projeto
